+++ components/home/HomeLanding.tsx
@@ -35,6 +35,7 @@ export default HomeLanding;
 const Styled = {
   Root: styled.div`
     width: 140.8rem;
+    min-height: 66rem;
     margin: 0 auto;
   `,
 };
