+++ hooks/queries/category.ts
@@ -1,7 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { deleteCategory, fetchAllCategory, fetchOneCategory, postCategory, updateCategory } from 'apis/category';
+import { deleteCategory, fetchAllCategory, postCategory, updateCategory } from 'apis/category';
 import { useRouter } from 'next/router';
-import { CreateCategory } from 'types/category/client';
 
 const QUERY_KEY = {
   category: 'category',
@@ -45,7 +44,6 @@ export const useDeleteCategory = (showToast: (message: string) => void) => {
       queryClient.invalidateQueries([QUERY_KEY.category]);
       if (router.asPath.startsWith('/category/')) {
         router.push('/category');
-        //showToast('삭제되었습니다.');
       }
       showToast('삭제되었습니다.');
     },
