+++ pages/home/index.tsx
@@ -1,13 +1,8 @@
-import useRedirect from '@/hooks/common/useRedirect';
 import Layout from 'components/common/Layout';
 import HomeLanding from 'components/home/HomeLanding';
 
 function Home() {
-  const { isLoading } = useRedirect();
-
-  return isLoading ? (
-    <></>
-  ) : (
+  return (
     <Layout>
       <HomeLanding />
     </Layout>
