+++ pages/landing/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { BigLogoImg, ChromeWebstoreImg } from 'assets/img';
 import Image from 'next/image';
 import styled from 'styled-components';
