+++ pages/login/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useGoogleLogin } from '@react-oauth/google';
 import { GoogleLoginImg, OwnSizeLogoImg } from 'assets/img';
 import axios from 'axios';
@@ -10,7 +10,6 @@ import theme from 'styles/theme';
 import { lottieMapper } from '@/components/common/modal/LottieModal';
 import LottiePlayer from '@/components/common/modal/LottiePlayer';
 import { useAuth } from '@/hooks/business/user';
-import useRedirect from '@/hooks/common/useRedirect';
 import Layout from 'components/common/Layout';
 
 function Login() {
@@ -39,28 +38,23 @@ function Login() {
       setPage((prev) => (prev === lottieMapper.length - 1 ? 0 : prev + 1));
     }
   };
-  const { isLoading } = useRedirect();
 
   return (
     <Layout noHeader noMenuBar noFooter>
-      {isLoading ? (
-        <></>
-      ) : (
-        <Styled.Root>
-          <Styled.Header>
-            <Image src={OwnSizeLogoImg} alt="로고 이미지" />
-          </Styled.Header>
-          <LottiePlayer page={page} onClickArrow={onClickArrow} lottie={lottieMapper[page].lottie} isDarkMode isTop />
-          <Styled.LoginButton onClick={() => login()}>
-            <Image src={GoogleLoginImg} alt="구글로그인 버튼 이미지" />
-          </Styled.LoginButton>
-          <Styled.Message>
-            로그인은 개인 정보 보호 정책 및 서비스 약관에 동의하는 것을 의미하며,
-            <br />
-            서비스 이용을 위해 이메일과 이름, 프로필 이미지를 수집합니다.
-          </Styled.Message>
-        </Styled.Root>
-      )}
+      <Styled.Root>
+        <Styled.Header>
+          <Image src={OwnSizeLogoImg} alt="로고 이미지" />
+        </Styled.Header>
+        <LottiePlayer page={page} onClickArrow={onClickArrow} lottie={lottieMapper[page].lottie} isDarkMode isTop />
+        <Styled.LoginButton onClick={() => login()}>
+          <Image src={GoogleLoginImg} alt="구글로그인 버튼 이미지" />
+        </Styled.LoginButton>
+        <Styled.Message>
+          로그인은 개인 정보 보호 정책 및 서비스 약관에 동의하는 것을 의미하며,
+          <br />
+          서비스 이용을 위해 이메일과 이름, 프로필 이미지를 수집합니다.
+        </Styled.Message>
+      </Styled.Root>
     </Layout>
   );
 }
@@ -92,7 +86,6 @@ const Styled = {
     margin-top: 3.8rem;
     border: 0;
     background: transparent;
-    cursor: pointer;
   `,
   Message: styled.h1`
     margin-top: 6.2rem;
