+++ pages/mysize/index.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 
-import useRedirect from '@/hooks/common/useRedirect';
 import Layout from 'components/common/Layout';
-import SizeForm from 'components/common/SizeForm/SizeForm';
 import Mysize from 'components/mysize/Mysize';
 
 export default function MySize() {
-  const { isLoading } = useRedirect();
   return (
     <Layout noMenuBar noFooter>
-      <Mysize></Mysize>
+      <Mysize />
     </Layout>
   );
 }
